+++ vkconfig_core/configurator.cpp
@@ -824,8 +824,17 @@ bool Configurator::Load() {
             const QJsonObject& json_object = json_interface_object.value(GetToken(TAB_PREFERENCES)).toObject();
 
             this->use_layer_dev_mode = json_object.value("use_layer_dev_mode").toBool();
+            this->latest_sdk_version = Version(json_object.value("latest_sdk_version").toString().toStdString().c_str());
+
+            if (json_object.value("use_notify_releases") != QJsonValue::Undefined) {
+                this->use_notify_releases = json_object.value("use_notify_releases").toBool();
+            }
+
             this->use_system_tray = json_object.value("use_system_tray").toBool();
             ::SetHomePath(json_object.value("VK_HOME").toString().toStdString());
+            if (json_object.value("VK_DOWNLOAD") != QJsonValue::Undefined) {
+                ::SetDownloadPath(json_object.value("VK_DOWNLOAD").toString().toStdString());
+            }
         }
 
         this->executables.Load(json_root_object);
@@ -873,7 +882,10 @@ bool Configurator::Save() const {
         QJsonObject json_object;
         json_object.insert("use_system_tray", this->use_system_tray);
         json_object.insert("use_layer_dev_mode", this->use_layer_dev_mode);
+        json_object.insert("use_notify_releases", this->use_notify_releases);
+        json_object.insert("latest_sdk_version", this->latest_sdk_version.str().c_str());
         json_object.insert("VK_HOME", ::Path(Path::HOME).RelativePath().c_str());
+        json_object.insert("VK_DOWNLOAD", ::Path(Path::DOWNLOAD).RelativePath().c_str());
         json_interface_object.insert(GetToken(TAB_PREFERENCES), json_object);
     }
 
@@ -932,6 +944,10 @@ void Configurator::SetUseSystemTray(bool enabled) { this->use_system_tray = enab
 bool Configurator::GetUseLayerDevMode() const { return this->use_layer_dev_mode; }
 
 void Configurator::SetUseLayerDevMode(bool enabled) { this->use_layer_dev_mode = enabled; }
+
+bool Configurator::GetUseNotifyReleases() const { return this->use_notify_releases; }
+
+void Configurator::SetUseNotifyReleases(bool enabled) { this->use_notify_releases = enabled; }
 
 bool Configurator::HasActiveSettings() const {
     switch (this->executable_scope) {