+++ vkconfig_gui/CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features:
 - Add dark mode support, automatically matching the OS mode
 - Add `vkconfig loader` commands to import, export, delete and use stored loader configurations
+- Add new Vulkan SDK notification and download
 
 ### Improvements:
 - Add button to remove missing layers